BACKEND/requirements.txt
fastapi==0.95.2
uvicorn==0.21.0
pandas
numpy
httpx==0.23.3
pydantic==1.10.12
python-multipart==0.0.5
openpyxl==3.1.2
xlrd==2.0.1

# LLM & Embeddings
ollama
langchain==0.0.327
huggingface-hub
sentence-transformers

# Vector Database
chromadb==0.4.14

# Agent Framework & Orchestration
pydantic-settings==2.0.3

# Utilities
python-dotenv
pyyaml
tenacity
typing-extensions

# Logging & Monitoring
python-json-logger

# Testing
pytest
pytest-asyncio
